Move spec into main crate
